+++ src/main/java/com/example/authserver/domain/entity/QREntity.java
@@ -16,13 +16,13 @@ import java.util.List;
         @StoredProcedureParameter(mode = ParameterMode.IN, name = "p_image",type = byte[].class),
         @StoredProcedureParameter(mode = ParameterMode.IN, name = "p_name",type = String.class),
         @StoredProcedureParameter(mode = ParameterMode.IN, name = "p_profile_id",type = Long.class),
-        @StoredProcedureParameter(mode = ParameterMode.OUT, name = "p_success",type = Integer.class),
+        @StoredProcedureParameter(mode = ParameterMode.OUT, name = "p_success",type = Long.class),
 })
 
 @NamedStoredProcedureQuery(name = "qrcode.deleteQR",procedureName = "DELETE_QRCODE",parameters = {
 
         @StoredProcedureParameter(mode = ParameterMode.IN, name = "p_id",type = Long.class),
-        @StoredProcedureParameter(mode = ParameterMode.OUT, name = "p_success",type = Integer.class),
+        @StoredProcedureParameter(mode = ParameterMode.OUT, name = "p_success",type = Long.class),
 })
 
 @NamedStoredProcedureQuery(name = "qrcode.updateQR",procedureName = "UPDATE_QRCODE",parameters = {
@@ -31,7 +31,7 @@ import java.util.List;
         @StoredProcedureParameter(mode = ParameterMode.IN, name = "p_description",type = String.class),
         @StoredProcedureParameter(mode = ParameterMode.IN, name = "p_name",type = String.class),
 
-        @StoredProcedureParameter(mode = ParameterMode.OUT, name = "p_success",type = Integer.class),
+        @StoredProcedureParameter(mode = ParameterMode.OUT, name = "p_success",type = Long.class),
 })
 public class QREntity {
     @Id
